requirements.txt
pymongo==4.0.2
python-dotenv==0.19.2
Faker
